Simultaneous access duration dates editing (GSI-1658) (#150)
* Initial implementation, probably still not perfect.
Extended the edit field component to allow for multiple inputs simultaneously.
Minor fixes in the site footer and the AR dialog.

* Fixed ticket URL applying to all fields

* Implemented suggestion
Implemented validation for dates
Made it so only the first item has the edit and cancel/save buttons

* Implemented suggestions
Moved date editing to its own component and restored single field editing for the field edit component.
Improved date validation in both duration editing and access requesting.
Importantly, the user's time zone, and its effects of access and displaying access duration needs to be dealt with in more detail, which is likely not possible in this PR.

* Implemented suggestions.
Improvements to code.
Fixes to the displaying of the end date.

* Refactor duration edit template and add clearer wording for duration of approved requests.

* Fix test setup for access request duration edit component to provide proper mock config object.

---------

Co-authored-by: Pascal Kraft <kraft.pascal@gmail.com>